+++ cumulus_library/databases/__init__.py
@@ -1,0 +1,16 @@
+from .athena import AthenaDatabaseBackend, AthenaParser
+from .base import DatabaseBackend, DatabaseCursor, DatabaseParser
+from .duckdb import DuckDatabaseBackend, DuckDbParser
+from .utils import create_db_backend, read_ndjson_dir
+
+__all__ = [
+    "AthenaDatabaseBackend",
+    "AthenaParser",
+    "DatabaseCursor",
+    "DatabaseBackend",
+    "DatabaseParser",
+    "DuckDatabaseBackend",
+    "DuckDbParser",
+    "create_db_backend",
+    "read_ndjson_dir",
+]
